pyproject.toml
[tool.poetry]
name = "flask-zmq"
version = "0.1.0"
description = ""
authors = ["erhe <eric.hermelin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^2.2.3"
pydantic = "^2.8.2"
pyzmq = "^26.2.0"
anyio = "^4.4.0"
gunicorn = "23.0.0"
uvloop = "^0.20.0"
httpx = "^0.27.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
